+++ tasks/goroutines.go
@@ -1,9 +1,5 @@
 package tasks
 
-import (
-	"fmt"
-)
-
 // implement task in main.go
 
 // Task: Print Numbers 1 to 10 Using Goroutines
@@ -15,15 +11,3 @@ import (
 // Use goroutines to ensure that the numbers are printed concurrently.
 
 // Print the numbers in ascending order with a 1-second delay between each number.
-
-func printFirstHalf() {
-	for i := 1; i <= 5; i++ {
-		fmt.Println(i)
-	}
-}
-
-func printSecondHalf() {
-	for i := 6; i <= 10; i++ {
-		fmt.Println(i)
-	}
-}
